fix: harden document verification pipeline

+++ app/interfaces/schemas/response_schemas.py
@@ -18,7 +18,7 @@ class JobResultResponse(BaseModel):
     queue_id: str
     ai_return_status: str
     ai_return_cd: str
-    ai_return_confidence: int | None
+    ai_return_confidence: float | None
     ai_return_remark: str
     result: dict[str, Any] | None = None
     pages: list[dict[str, Any]] = Field(default_factory=list, description="Per-page OCR, detections, barcode")
@@ -28,7 +28,6 @@ class ReprocessResponse(BaseModel):
     queue_id: str
     status: str = "REPROCESS_QUEUED"
     message: str = "Job reprocess has been queued."
-
